Implement dynamodb schema repository
Writes/read of the actual schema are all made with full consistency; this
makes it much simpler to understand when schema changes occur.
Dynamic columns, on the other hand, are eventually consistent reads for two reasons:
(1) its ok to get a slightly older answer, and (2) the changes are expected
to be much higher frequency, which would be unnecessarily constricted by
locking on the fully consistent view.

There are two distinctly different row type:
 a. Schema specficiation
 b. Schema dynamic columns

For (a), the key is |subjectname|subjectname| and fields are
the configuration and evolving schema.

For (b) the key is |subjectname|subjectname_ext| and fields are
the column names and suggested column types.

This format ensures that subjects and their extensions are stored together,
but don't conflict when reading or locking on writes.